requirement.txt
Python==3.8
PyTorch==2.4.1
numpy==1.24.4         
scipy==1.10.1         
opencv-python==4.8.0.76
pillow==9.5.0        
scikit-learn==1.2.2   
matplotlib==3.7.5           
openai-clip==1.0
tqdm==4.66.4    